requirements.txt
langchain-openai
langchian-core
python-dotenv
streamlit
langchain-community
flask
